requirements.txt
# Market Mood Ring - Phase 1 Requirements (Default)
# ETL Pipeline with Sentiment Analysis
# For Phase 2 (RAG), see requirements-phase2.txt

# Core dependencies
requests>=2.31.0
pandas>=2.0.0

# Database
psycopg2-binary>=2.9.9

# Kafka
kafka-python>=2.0.2

# Dashboard
streamlit>=1.28.0
plotly>=5.17.0

# Phase 2 dependencies (RAG/AI) - NOT included by default
# sentence-transformers>=2.2.2  # Uncomment for Phase 2 (requires torch ~888MB)
